add common util
